Add mlock when reading symbols in index reader
