+++ tsdb/index/index.go
@@ -1082,6 +1082,11 @@ type Reader struct {
 	dec *Decoder
 
 	version int
+
+	// Function to read symbols
+	readSymbolsFunc func() (*Symbols, error)
+	// Function to close Reader
+	closeReaderFunc func() error
 }
 
 type postingOffset struct {
@@ -1132,13 +1137,7 @@ func NewFileReader(path string) (*Reader, error) {
 	return r, nil
 }
 
-func newReader(b ByteSlice, c io.Closer) (*Reader, error) {
-	r := &Reader{
-		b:        b,
-		c:        c,
-		postings: map[string][]postingOffset{},
-	}
-
+func createReader(r *Reader) (*Reader, error) {
 	// Verify header.
 	if r.b.Len() < HeaderLen {
 		return nil, errors.Wrap(encoding.ErrInvalidSize, "index header")
@@ -1153,12 +1152,12 @@ func newReader(b ByteSlice, c io.Closer) (*Reader, error) {
 	}
 
 	var err error
-	r.toc, err = NewTOCFromByteSlice(b)
+	r.toc, err = NewTOCFromByteSlice(r.b)
 	if err != nil {
 		return nil, errors.Wrap(err, "read TOC")
 	}
 
-	r.symbols, err = NewSymbols(r.b, r.version, int(r.toc.Symbols))
+	r.symbols, err = r.readSymbolsFunc()
 	if err != nil {
 		return nil, errors.Wrap(err, "read symbols")
 	}
@@ -1437,7 +1436,7 @@ func ReadPostingsOffsetTable(bs ByteSlice, off uint64, f func(name, value []byte
 
 // Close the reader and its underlying resources.
 func (r *Reader) Close() error {
-	return r.c.Close()
+	return r.closeReaderFunc()
 }
 
 func (r *Reader) lookupSymbol(o uint32) (string, error) {
