+++ tsdb/index/reader_builder_windows.go
@@ -1,0 +1,36 @@
+// Copyright 2017 The Prometheus Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build windows
+// +build windows
+
+package index
+
+func newReader(b ByteSlice, c io.Closer) (*Reader, error) {
+	r := &Reader{
+		b:        b,
+		c:        c,
+		postings: map[string][]postingOffset{},
+	}
+	r.readSymbolsFunc = func() (*Symbols, error) {
+		symbols, err = NewSymbols(r.b, r.version, int(r.toc.Symbols))
+		if err != nil {
+			return nil, errors.Wrap(err, "read symbols")
+		}
+		return symbols, nil
+	}
+	r.closeReaderFunc = func() error {
+		return r.c.Close()
+	}
+	return createReader(r)
+}
